search页面初步完成，还需点击事件

+++ module_search/src/main/java/com/example/module/search/bean/MusicData.kt
@@ -11,18 +11,18 @@ data class ResultMusic(
 data class Song(
     val album: Album,
     val artists: List<ArtistX>,
-    val id: Int,
-    val mvid: Int,
+    val id: Long,
+    val mvid: Long,
     val name: String,
 )
 
 data class Album(
-    val id: Int,
+    val id: Long,
     val name: String,
 )
 
 data class ArtistX(
-    val id: Int,
+    val id: Long,
     val name: String,
     val picUrl: Any,
 )
